compilation warning removed
git-svn-id: svn://svn.code.sf.net/p/hxcfloppyemu/code/HxCFloppyEmulator/HxCFloppyEmulator_software/trunk@612 63fa2f70-6a9b-4bc4-b855-9c6a949b1d69

+++ HxCFloppyEmulator_software/sources/loader.cxx
@@ -177,7 +177,7 @@ int loadrawfile(HXCFLOPPYEMULATOR* floppycontext,cfgrawfile * rfc,char * file)
 			hxcfe_setTrackInterleave(fb,rfc->interleave);	
 			hxcfe_setSectorFill(fb,rfc->fillvalue);
 
-			hxcfe_setTrackPreGap(fb, rfc->pregap);
+			hxcfe_setTrackPreGap(fb, (unsigned short)rfc->pregap);
 
 			if(rfc->autogap3)
 				hxcfe_setSectorGap3(fb,255);
